Переносим в презентер рендеринг
